# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Subframe Protocol API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type SubdomainStatus =
  (typeof SubdomainStatus)[keyof typeof SubdomainStatus];

export const SubdomainStatus = {
  pending: "pending",
  active: "active",
  linked: "linked",
} as const;

export type SubdomainTokenStatus =
  (typeof SubdomainTokenStatus)[keyof typeof SubdomainTokenStatus];

export const SubdomainTokenStatus = {
  none: "none",
  deploying: "deploying",
  deployed: "deployed",
  failed: "failed",
} as const;

export interface Subdomain {
  id: number;
  name: string;
  walletAddress: string;
  ensFullName: string;
  /** @nullable */
  ipfsCid: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bio: string | null;
  status: SubdomainStatus;
  /** @nullable */
  ensTx1Hash: string | null;
  /** @nullable */
  ensTx2Hash: string | null;
  /** @nullable */
  ensTx3Hash: string | null;
  /** @nullable */
  ensTx4Hash: string | null;
  tokenStatus: SubdomainTokenStatus;
  /** @nullable */
  tokenAddress: string | null;
  /** @nullable */
  tokenSymbol: string | null;
  /** @nullable */
  tokenName: string | null;
  /** @nullable */
  tokenDeployTxHash: string | null;
  /** @nullable */
  uniswapPairAddress: string | null;
  /** @nullable */
  uniswapLiquidityTxHash: string | null;
  /** @nullable */
  artTokenId: string | null;
  claimedAt: string;
  updatedAt: string;
}

export interface CreateSubdomainBody {
  name: string;
  walletAddress: string;
  bio?: string;
  avatarUrl?: string;
}

export type UpdateSubdomainBodyStatus =
  (typeof UpdateSubdomainBodyStatus)[keyof typeof UpdateSubdomainBodyStatus];

export const UpdateSubdomainBodyStatus = {
  pending: "pending",
  active: "active",
  linked: "linked",
} as const;

export interface UpdateSubdomainBody {
  ipfsCid?: string;
  bio?: string;
  avatarUrl?: string;
  status?: UpdateSubdomainBodyStatus;
}

export interface AvailabilityResult {
  available: boolean;
  name: string;
}

export interface ProtocolStats {
  totalSubdomains: number;
  activeSubdomains: number;
  linkedToIPFS: number;
  recentClaims: Subdomain[];
}

export interface Transaction {
  hash: string;
  from: string;
  /** @nullable */
  to: string | null;
  value: string;
  valueEth: string;
  gasUsed: string;
  timestamp: number;
  /** @nullable */
  method: string | null;
  status: string;
}

export interface WalletData {
  address: string;
  /** @nullable */
  ensName: string | null;
  balanceEth: string;
  /** @nullable */
  balanceUsd: string | null;
  txCount: number;
  lastTransactions: Transaction[];
  /** @nullable */
  firstSeen: string | null;
  isContract: boolean;
}

export type WalletAnalysisRiskLevel =
  (typeof WalletAnalysisRiskLevel)[keyof typeof WalletAnalysisRiskLevel];

export const WalletAnalysisRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface WalletAnalysis {
  address: string;
  summary: string;
  activityType: string;
  riskLevel: WalletAnalysisRiskLevel;
  tags: string[];
  insights: string[];
  generatedAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
  walletContext?: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}