# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Subframe Protocol API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subdomains:
    get:
      operationId: listSubdomains
      tags: [subdomains]
      summary: List all registered subdomains
      responses:
        "200":
          description: List of registered subdomains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subdomain"
    post:
      operationId: createSubdomain
      tags: [subdomains]
      summary: Register a new subdomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubdomainBody"
      responses:
        "201":
          description: Created subdomain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "409":
          description: Name already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /subdomains/check/{name}:
    get:
      operationId: checkSubdomainAvailability
      tags: [subdomains]
      summary: Check if a subdomain name is available
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResult"
  /subdomains/{id}:
    get:
      operationId: getSubdomain
      tags: [subdomains]
      summary: Get a subdomain by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subdomain details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateSubdomain
      tags: [subdomains]
      summary: Update a subdomain (IPFS CID, avatar, bio)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubdomainBody"
      responses:
        "200":
          description: Updated subdomain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /subdomains/by-name/{name}:
    get:
      operationId: getSubdomainByName
      tags: [subdomains]
      summary: Get subdomain by ENS name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subdomain details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Subdomain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        walletAddress:
          type: string
        ensFullName:
          type: string
        ipfsCid:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, linked]
        ensTx1Hash:
          type: ["string", "null"]
        ensTx2Hash:
          type: ["string", "null"]
        ensTx3Hash:
          type: ["string", "null"]
        ensTx4Hash:
          type: ["string", "null"]
        tokenStatus:
          type: string
          enum: [none, deploying, deployed, failed]
        tokenAddress:
          type: ["string", "null"]
        tokenSymbol:
          type: ["string", "null"]
        tokenName:
          type: ["string", "null"]
        tokenDeployTxHash:
          type: ["string", "null"]
        uniswapPairAddress:
          type: ["string", "null"]
        uniswapLiquidityTxHash:
          type: ["string", "null"]
        artTokenId:
          type: ["string", "null"]
        claimedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - walletAddress
        - ensFullName
        - ipfsCid
        - avatarUrl
        - bio
        - status
        - ensTx1Hash
        - ensTx2Hash
        - ensTx3Hash
        - ensTx4Hash
        - tokenStatus
        - tokenAddress
        - tokenSymbol
        - tokenName
        - tokenDeployTxHash
        - uniswapPairAddress
        - uniswapLiquidityTxHash
        - artTokenId
        - claimedAt
        - updatedAt
    CreateSubdomainBody:
      type: object
      properties:
        name:
          type: string
        walletAddress:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
      required:
        - name
        - walletAddress
    UpdateSubdomainBody:
      type: object
      properties:
        ipfsCid:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
          enum: [pending, active, linked]
    AvailabilityResult:
      type: object
      properties:
        available:
          type: boolean
        name:
          type: string
      required:
        - available
        - name
...